implemented some of Yohann's review
